+++ packages/core-data/src/utils/crdt.ts
@@ -21,6 +21,7 @@ import {
 import { BaseAwareness } from '../awareness/base-awareness';
 import {
 	mergeCrdtBlocks,
+	mergeRichTextUpdate,
 	type Block,
 	type YBlock,
 	type YBlocks,
@@ -45,6 +46,7 @@ import {
 // Changes that can be applied to a post entity record.
 export type PostChanges = Partial< Post > & {
 	blocks?: Block[];
+	content?: Post[ 'content' ] | string;
 	excerpt?: Post[ 'excerpt' ] | string;
 	selection?: WPSelection;
 	title?: Post[ 'title' ] | string;
@@ -53,11 +55,13 @@ export type PostChanges = Partial< Post > & {
 // A post record as represented in the CRDT document (Y.Map).
 export interface YPostRecord extends YMapRecord {
 	author: number;
-	blocks: YBlocks;
+	// Blocks are undefined when they need to be re-parsed from content.
+	blocks: YBlocks | undefined;
+	content: Y.Text;
 	categories: number[];
 	comment_status: string;
 	date: string | null;
-	excerpt: string;
+	excerpt: Y.Text;
 	featured_media: number;
 	format: string;
 	meta: YMapWrap< YMapRecord >;
@@ -67,13 +71,14 @@ export interface YPostRecord extends YMapRecord {
 	sticky: boolean;
 	tags: number[];
 	template: string;
-	title: string;
+	title: Y.Text;
 }
 
 // Properties that are allowed to be synced for a post.
 const allowedPostProperties = new Set< string >( [
 	'author',
 	'blocks',
+	'content',
 	'categories',
 	'comment_status',
 	'date',
@@ -156,6 +161,14 @@ export function applyPostChangesToCRDTDoc(
 
 		switch ( key ) {
 			case 'blocks': {
+				// Blocks are undefined when they need to be re-parsed from content.
+				if ( ! newValue ) {
+					// Set to undefined instead of deleting the key. This is important
+					// since we iterate over the Y.Map keys in getPostChangesFromCRDTDoc.
+					ymap.set( key, undefined );
+					break;
+				}
+
 				let currentBlocks = ymap.get( key );
 
 				// Initialize.
@@ -164,9 +177,6 @@ export function applyPostChangesToCRDTDoc(
 					ymap.set( key, currentBlocks );
 				}
 
-				// Block[] from local changes.
-				const newBlocks = ( newValue as PostChanges[ 'blocks' ] ) ?? [];
-
 				// Block changes from typing are bundled with a 'selection' update.
 				// Pass the resulting cursor position to the mergeCrdtBlocks function.
 				const cursorPosition =
@@ -174,15 +184,33 @@ export function applyPostChangesToCRDTDoc(
 
 				// Merge blocks does not need `setValue` because it is operating on a
 				// Yjs type that is already in the Y.Doc.
-				mergeCrdtBlocks( currentBlocks, newBlocks, cursorPosition );
+				mergeCrdtBlocks( currentBlocks, newValue, cursorPosition );
 				break;
 			}
 
-			case 'excerpt': {
-				const currentValue = ymap.get( 'excerpt' );
-				const rawNewValue = getRawValue( newValue );
+			case 'content':
+			case 'excerpt':
+			case 'title': {
+				const currentValue = ymap.get( key );
+				let rawValue = getRawValue( newValue );
 
-				updateMapValue( ymap, key, currentValue, rawNewValue );
+				// Copy logic from prePersistPostType to ensure that the "Auto
+				// Draft" template title is not synced.
+				if (
+					key === 'title' &&
+					! currentValue &&
+					'Auto Draft' === rawValue
+				) {
+					rawValue = '';
+				}
+
+				if ( currentValue instanceof Y.Text ) {
+					mergeRichTextUpdate( currentValue, rawValue ?? '' );
+				} else {
+					const newYText = new Y.Text( rawValue ?? '' );
+					ymap.set( key, newYText );
+				}
+
 				break;
 			}
 
@@ -224,20 +252,6 @@ export function applyPostChangesToCRDTDoc(
 
 				const currentValue = ymap.get( key );
 				updateMapValue( ymap, key, currentValue, newValue );
-				break;
-			}
-
-			case 'title': {
-				const currentValue = ymap.get( key );
-
-				// Copy logic from prePersistPostType to ensure that the "Auto
-				// Draft" template title is not synced.
-				let rawNewValue = getRawValue( newValue );
-				if ( ! currentValue && 'Auto Draft' === rawNewValue ) {
-					rawNewValue = '';
-				}
-
-				updateMapValue( ymap, key, currentValue, rawNewValue );
 				break;
 			}
 
@@ -318,11 +332,11 @@ export function getPostChangesFromCRDTDoc(
 						ydoc.meta?.get( CRDT_DOC_META_PERSISTENCE_KEY ) &&
 						editedRecord.content
 					) {
-						const blocks = ymap.get( 'blocks' ) as YBlocks;
+						const blocksJson = ymap.get( 'blocks' )?.toJSON() ?? [];
+
 						return (
-							__unstableSerializeAndClean(
-								blocks.toJSON()
-							).trim() !== editedRecord.content.raw.trim()
+							__unstableSerializeAndClean( blocksJson ).trim() !==
+							getRawValue( editedRecord.content )
 						);
 					}
 
@@ -375,6 +389,7 @@ export function getPostChangesFromCRDTDoc(
 					return haveValuesChanged( currentValue, newValue );
 				}
 
+				case 'content':
 				case 'excerpt':
 				case 'title': {
 					return haveValuesChanged(