+++ packages/edit-site/src/components/page-templates/hooks.js
@@ -22,14 +22,14 @@ import { TEMPLATE_ORIGINS } from '../../utils/constants';
  *
  * @typedef AddedByData
  * @type {Object}
- * @property {AddedByType}  type         The type of the data.
- * @property {JSX.Element}  icon         The icon to display.
- * @property {string}       [imageUrl]   The optional image URL to display.
- * @property {string}       [text]       The text to display.
- * @property {boolean}      isCustomized Whether the template has been customized.
+ * @property {AddedByType}       type         The type of the data.
+ * @property {React.JSX.Element} icon         The icon to display.
+ * @property {string}            [imageUrl]   The optional image URL to display.
+ * @property {string}            [text]       The text to display.
+ * @property {boolean}           isCustomized Whether the template has been customized.
  *
- * @param    {TemplateType} postType     The template post type.
- * @param    {number}       postId       The template post id.
+ * @param    {TemplateType}      postType     The template post type.
+ * @param    {number}            postId       The template post id.
  * @return {AddedByData} The added by object or null.
  */
 export function useAddedBy( postType, postId ) {